% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotcALB-package.R
\docType{package}
\name{iotcALB-package}
\alias{iotcALB}
\alias{iotcALB-package}
\title{iotcALB: Indian Ocean Albacore Management Strategy Evaluation}
\description{
Operating models, management procedures, and analysis tools for the Management Strategy Evaluation (MSE) of Indian Ocean albacore tuna.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Blue-Matter/iotcALB}
  \item \url{https://blue-matter.github.io/iotcALB/}
  \item Report bugs at \url{https://github.com/Blue-Matter/iotcALB/issues}
}

}
\author{
\strong{Maintainer}: Adrian Hordyk \email{adrian@bluematterscience.com}

Authors:
\itemize{
  \item Adrian Hordyk \email{adrian@bluematterscience.com}
}

}
\keyword{internal}
